requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-filter==22.1
django-json==0.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
zipp==3.15.0
